+++ blocks/header/header.css
@@ -5,8 +5,6 @@ header nav {
     'hamburger brand tools' var(--nav-height)
     'sections sections sections' 1fr / 50px 1fr 50px;
   align-items: center;
-  padding-top: 50px;
-  padding-bottom: 50px;
   position: fixed;
   z-index: 1;
   background-color: var(--background-color);
